# backend/api/cobytes-openapi.yml
openapi: 3.0.3
info:
  title: Cobytes Security Platform API
  description: |
    The Cobytes Security Platform API provides comprehensive security scanning and reporting capabilities.
    
    ## Authentication
    All endpoints (except health and API info) require JWT authentication.
    Use the `/api/auth/login` endpoint to obtain a token.
    
    ## Base URL
    - Development: `http://localhost:3001`
    - Production: `https://api.cobytes.com`
  version: 1.0.0
  contact:
    name: Cobytes API Support
    email: api-support@cobytes.com
    url: https://docs.cobytes.com
  license:
    name: Proprietary
    url: https://cobytes.com/license
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.cobytes.com
    description: Production server
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Authenticate a user and receive a JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register
      description: Register a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/verify:
    get:
      tags:
        - Authentication
      summary: Verify Token
      description: Verify the validity of a JWT token
      operationId: verifyToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh Token
      description: Refresh an existing JWT token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout
      description: Logout the current user
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
  /api/scans:
    get:
      tags:
        - Scans
      summary: Get All Scans
      description: Get a list of all scans for the authenticated user
      operationId: getAllScans
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by scan status
          schema:
            type: string
            enum: [pending, running, completed, failed]
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of scans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScansListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Scans
      summary: Create New Scan
      description: Create and start a new security scan
      operationId: createScan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScanRequest'
      responses:
        '201':
          description: Scan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/reports:
    get:
      tags:
        - Reports
      summary: Get All Reports
      description: Get a list of all generated reports
      operationId: getAllReports
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@cobytes.com
        password:
          type: string
          format: password
          example: admin123
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: securePassword123
        name:
          type: string
          example: John Doe
        organization:
          type: string
          example: Example Corp
    User:
      type: object
      properties:
        id:
          type: string
          example: user_1
        email:
          type: string
          format: email
          example: admin@cobytes.com
        role:
          type: string
          enum: [user, admin]
          example: admin
        name:
          type: string
          example: Admin User
        organization:
          type: string
          example: Cobytes Inc
        createdAt:
          type: string
          format: date-time
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIs...
            expiresIn:
              type: integer
              example: 86400
    VerifyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: object
              properties:
                valid:
                  type: boolean
                  example: true
                expiresAt:
                  type: string
                  format: date-time
    RefreshRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
    RefreshResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIs...
            expiresIn:
              type: integer
              example: 86400
    LogoutResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Logged out successfully
    CreateScanRequest:
      type: object
      required:
        - scanType
        - targets
      properties:
        scanType:
          type: string
          enum: [web_vulnerability, network_scan, ssl_tls, dns_security, malware_analysis]
          example: web_vulnerability
        targets:
          type: array
          items:
            type: string
          example: [https://example.com, https://test.example.com]
        configuration:
          type: object
          properties:
            depth:
              type: integer
              example: 3
            threads:
              type: integer
              example: 10
            followRedirects:
              type: boolean
              example: true
            userAgent:
              type: string
              example: Cobytes Security Scanner
        scheduledAt:
          type: string
          format: date-time
    Scan:
      type: object
      properties:
        scanId:
          type: string
          example: scan_1748799588916_dw04msszk
        status:
          type: string
          enum: [pending, running, completed, failed]
          example: pending
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScanResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Scan'
    ScansListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Scan'
        count:
          type: integer
          example: 1
...
    Report:
      type: object
      properties:
        reportId:
          type: string
          example: report_1748799588923_9rehl4b0k
        name:
          type: string
          example: Security Report 2025-06-01
        scanIds:
          type: array
          items:
            type: string
        format:
          type: string
          enum: [pdf, html, json, csv]
          example: pdf
        status:
          type: string
          enum: [generating, completed, failed]
          example: generating
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        size:
          type: integer
          example: 245760
        downloadUrl:
          type: string
          nullable: true
          example: /api/reports/report_1748799588923_9rehl4b0k/download
...
    ReportsListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Report'
        count:
          type: integer
          example: 1
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        code:
          type: string
          example: ERROR_CODE
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              field: Error details
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Invalid or missing authentication
            code: UNAUTHORIZED
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Resource already exists
            code: CONFLICT
...